backend/requirements.txt
django~=2.2
django-filter~=2.3
djangorestframework~=3.12
djangorestframework-simplejwt~=4.8
djoser~=2.1
gunicorn~=20.0
psycopg2-binary==2.8.6
PyJWT~=2.1
pytz~=2020.1
requests~=2.26
python-dotenv~=0.21
Pillow~=8.3
django-colorfield~=0.8
drf_extra_fields